+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "classroom-assignment" extension will be documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [1.0.0] - 2025-11-23
 
-- Initial release
+
+
+- Initial release of the "GitHub Classroom Assignment Viewer" extension.
+
+- Added functionality to display reports, grades, and attempt numbers from GitHub Actions workflow run artifacts for student repositories.